test: making it so my tools are somewhat isolated from the system in both regular use and within devcontainers

+++ scripts/bootstrap.sh
@@ -1,25 +1,43 @@
 #!/usr/bin/env bash
 
-### this script is ONLY for Bootstrapping dev container environments
-### that are using a standard devcontainer.json file and with some
-### tool like devpod
+# This script is for bootstrapping environments for neovim use
+# Works as a way to package all the tooling I need for
+# neovim without needing root access to the system
 
 set -e
 
-DOTFILES="$HOME/dotfiles"
+# main boolean to check if nothing changed or not
+CHANGED=false
+
+# check if not container to set dotfiles dir
+if [ ! -f "/.dockerenv" ]; then
+    DOTFILES="$HOME/dotfiles"
+else
+    DOTFILES="$HOME/.dotfiles"
+fi
+
 USERBIN="$HOME/.local/bin"
 MISEBIN="$USERBIN/mise"
 
-# creates .config if not already
+# creates ~/.config if not already there
 if [ ! -d "$HOME/.config" ]; then
+    CHANGED=true
     echo "Creating .config dir..."
     mkdir -p "$HOME/.config"
 fi
 
-# creates .local/bin for binaries
+# creates ~/.local/bin if not already there
 if [ ! -d "$HOME/.local/bin" ]; then
+    CHANGED=true
     echo "Creating .local/bin dir..."
     mkdir -p "$HOME/.local/bin"
+fi
+
+# creates ~/.bashrc.d if not already there
+if [ ! -d "$HOME/.bashrc.d" ]; then
+    CHANGED=true
+    echo "Creating .bashrc.d dir..."
+    mkdir -p "$HOME/.bashrc.d"
 fi
 
 # function that makes dotfiles link easier
@@ -27,59 +45,84 @@ dotlink() {
     local dotLocation="$DOTFILES/$1"
     local confLocation="$HOME/$2"
 
-    echo "linking $dotLocation to $confLocation"
-    ln -sfn "$dotLocation" "$confLocation"
+    if [ ! -d "$confLocation" ] && [ ! -f "$confLocation" ]; then
+        CHANGED=true
+        echo "Linking [$dotLocation] to [$confLocation]"
+        ln -sfn "$dotLocation" "$confLocation"
+    fi
 }
 
-echo "Linking dotfiles to host..."
-
 # Links files downloaded from github to user environment config locations
-dotlink "bash/.bashrc.d" ".bashrc.d"
+dotlink "bash/.bashrc.d/prompt.sh" ".bashrc.d/prompt.sh"
 dotlink "nvim" ".config/nvim"
 dotlink "tmux/.tmux.conf" ".tmux.conf"
 
-echo "Dotfiles linked to config dirs"
-echo "Appending source spript to .bashrc..."
+USERPATH="
+#--- sets user path to .local/bin
+export PATH=\"\$PATH:\$HOME/.local/bin\"
+#--- end of setting user path
+"
+
+# if user path isn't added already
+if ! grep -q "sets user path to" "$HOME/.bashrc"; then
+    CHANGED=true
+    echo "Adding user path to .bashrc..."
+    echo "$USERPATH" >> "$HOME/.bashrc"
+fi
 
 # download mise if not already on the system
 if [ ! -f "$MISEBIN" ]; then
+    CHANGED=true
+    echo "Downloading mise via curl..."
     curl -Lo "$MISEBIN" "https://mise.jdx.dev/mise-latest-linux-x64" 2>&1
     chmod +x "$MISEBIN"
 fi
 
-ACTIVATE="
-#--- activate mise to allow its tools to be in PATH
-
-eval \"\$(\$HOME/.local/bin/mise activate bash)\"
-
-#--- end of mise activation
+MISEPATH="
+#--- sets mise tool shims to be in path
+export PATH=\$PATH:\"\$HOME/.local/share/mise/shims\"
+#--- end of mise shims config
 "
 
-# if mise isn't activated yet then add to .bashrc
-# uses made up code to determine it
-if ! grep -q "MISE:146325" "$HOME/.bashrc"; then
-    echo "$ACTIVATE" >> "$HOME/.bashrc"
+# only do anything with mise if the project is inside a container
+# and if the env var is not set yet
+if [ ! -f "/.dockerenv" ]; then
+    if ! grep -q "mise tool shims" "$HOME/.bashrc"; then
+        CHANGED=true
+        echo "Adding mise tools to PATH"
+        echo "$MISEPATH" >> "$HOME/.bashrc"
+    fi
 fi
+
+checkmise() {
+    local name=$1
+    local version=$2
+    local tool=$name@$version
+
+    if ! mise where "$tool" >/dev/null 2>&1; then
+	$MISEBIN use -g "$tool"
+	CHANGED=true
+    fi
+}
 
 # sets specific versions to use
 NEOVIM_VERSION="0.11.6"
 TMUX_VERSION="3.6a"
-RIPGREP_VERSION="15.1.0"
+RG_VERSION="15.1.0"
 FD_VERSION="10.3.0"
 FZF_VERSION="0.67.0"
 PY_VERSION="3.14.3"
 NODE_VERSION="25.6.1"
 
-# installs tooling with mise in path dir
-$MISEBIN use -g neovim@$NEOVIM_VERSION
-$MISEBIN use -g tmux@$TMUX_VERSION
-$MISEBIN use -g ripgrep@$RIPGREP_VERSION
-$MISEBIN use -g fd@$FD_VERSION
-$MISEBIN use -g fzf@$FZF_VERSION
-$MISEBIN use -g python@$PY_VERSION
-$MISEBIN use -g node@$NODE_VERSION
+checkmise "neovim" "$NEOVIM_VERSION"
+checkmise "tmux" "$TMUX_VERSION"
+checkmise "ripgrep" "$RG_VERSION"
+checkmise "fd" "$FD_VERSION"
+checkmise "fzf" "$FZF_VERSION"
+checkmise "python" "$PY_VERSION"
+checkmise "node" "$NODE_VERSION"
 
-SCRIPT="
+SOURCE="
 # --- start of dotfiles config link ---
 if [ -d \"\$HOME/.bashrc.d\" ]; then
     for file in \"\$HOME/.bashrc.d/\"*; do
@@ -87,7 +130,18 @@ if [ -d \"\$HOME/.bashrc.d\" ]; then
     done
 fi
 unset file
+# --- end of dotfile config link ---
+"
 
+# check if .bashrc.d isn't mentioned in .bashrc
+if ! grep -q "start of dotfiles config link" "$HOME/.bashrc" && ! grep -q ".bashrc.d" "$HOME/.bashrc"; then
+    CHANGED=true
+    echo "Adding .bashrc.d sourcing to .bashrc..."
+    echo "$SOURCE" >> "$HOME/.bashrc"
+fi
+
+TMUX="
+# --- start of devcontainer tmux config ---
 # only run in interactive shells
 if [[ \$- == *i* ]]; then
 
@@ -100,11 +154,19 @@ if [[ \$- == *i* ]]; then
         exit 0
     fi
 fi
-
-# --- end of dotfile config link ---
+# --- end of devcontainer tmux config ---
 "
 
-# append the script to .bashrc
-echo "$SCRIPT" >> "$HOME/.bashrc"
+# check if host is a container
+if [ -f "/.dockerenv" ] && ! grep -q "start of devcontainer tmux config"; then
+    CHANGED=true
+    echo "Adding tmux sesssion auto attach to .bashrc..."
+    echo "$TMUX" >> "$HOME/.bashrc"
+fi
 
-echo "Bootstrapping finished"
+# output different based on CHANGED value
+if [ "$CHANGED" = true ]; then
+    echo "Bootstrapping finished"
+else
+    echo "System is already bootstrapped"
+fi
